Update run.sh
Corrected error s6-overlay-suexec: fatal. It can only run as pid 1 error typically means that the script inside the container is not running as the main process (PID 1). This happens because Home Assistant add-ons use the s6-overlay system for process management, and it expects the main process to be correctly structured.